feat: statusbar menu support
