+++ ReactNativeGUITools/Controller/MainProjectListVC/MainProjectListVC+DataSource.swift
@@ -9,6 +9,6 @@ import Cocoa
 
 extension MainProjectListVC: NSTableViewDataSource {
     func numberOfRows(in tableView: NSTableView) -> Int {
-        return self.projectInfoCollection.projectInfos.count
+        return ProjectInfoCollection.shared.projectInfos.count
     }
 }
