+++ ReactNativeGUITools/Controller/MainProjectListVC/MainProjectListVC+SearchDelegate.swift
@@ -10,19 +10,19 @@ import Cocoa
 extension MainProjectListVC: NSSearchFieldDelegate {
     func searchFieldDidStartSearching(_ sender: NSSearchField) {
         let searchString = sender.stringValue
-        projectInfoCollection.searchBy(searchString)
+        ProjectInfoCollection.shared.searchBy(searchString)
         projectListTableView.reloadData()
     }
 
     func searchFieldDidEndSearching(_ sender: NSSearchField) {
-        projectInfoCollection.clearSearch()
+        ProjectInfoCollection.shared.clearSearch()
         projectListTableView.reloadData()
     }
 
     func controlTextDidChange(_ obj: Notification) {
         guard let searchField = obj.object as? NSSearchField else { return }
         let searchString = searchField.stringValue
-        projectInfoCollection.searchBy(searchString)
+        ProjectInfoCollection.shared.searchBy(searchString)
         projectListTableView.reloadData()
     }
 }
